vision_helper/vision_helper.py
from google.cloud import vision
import io

def ocr(path):
    client = vision.ImageAnnotatorClient()
    with io.open(path, 'rb') as image_file:
        content = image_file.read()
    image = vision.types.Image(content=content)
    response = client.text_detection(image=image)
    if response.error.message:
        raise Exception(
            '{}\nFor more info on error messages, check: '
            'https://cloud.google.com/apis/design/errors'.format(
                response.error.message))
    annotates = response.text_annotations
    texts = []
    for annotate in annotates:
        texts.append(annotate.description)
    return texts