requirements.txt
Flask==1.1.2
line-bot-sdk
redis
pymongo
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-vision
grpcio